frontend: Notifications: Refactor notifications folder to Barrel Export Pattern
