+++ frontend/src/components/App/Notifications/List/index.tsx
@@ -1,0 +1,2 @@
+import NotificationList from './List';
+export default NotificationList; // for backword compatibility
